Allow drm-kmod to build with LINT.
Hide various #define DEBUG statements under #ifndef DEBUG to not re-define
DEBUG, or make sure DEBUG is not defined before we define DEBUG(...).

Also make sure KTR_DRM is always defined and fix a typo in one CTR4().

Further extend the size of a variable to check against a larger constant
value (which otherwise would always fail in error).

With this drm-kmod also builds along LINT kernels, which makes it easier
to test other changes.

Sponsored-by:	The FreeBSD Foundation
(cherry picked from commit b8a85669f61ffea54860a76f72b8c5c0bb647127)

+++ drivers/gpu/drm/i915/gem/i915_gem_execbuffer.c
@@ -1489,7 +1489,7 @@ static int eb_relocate_vma(struct i915_execbuffer *eb, struct i915_vma *vma)
 	struct drm_i915_gem_relocation_entry stack[N_RELOC(512)];
 	struct drm_i915_gem_relocation_entry __user *urelocs;
 	const struct drm_i915_gem_exec_object2 *entry = exec_entry(eb, vma);
-	unsigned int remain;
+	uintmax_t remain;
 
 	urelocs = u64_to_user_ptr(entry->relocs_ptr);
 	remain = entry->relocation_count;
